Massive update on the project
Had an issue with eclipse had to restart. Here is what this has done

Added an abstract factory that creates the cars needed and such

Created an employee class that had a strategy pattern built into the class for pay strcutre

Added an observer class for the employees when the pay structure is modified each pay strucutre is a seperate class for ease of use.